+++ checkpoint/orbax/checkpoint/_src/serialization/serialization.py
@@ -40,12 +40,6 @@ import tensorstore as ts
 TS_CONTEXT = ts.Context({'file_io_concurrency': {'limit': 128}})
 _REMOVED_VALUE = 'Value removed'
 _CHECKPOINT_SUCCESS = 'checkpoint_write_success'
-_REMOTE_URL_PREFIXES = ['gs://', 's3://']
-_REMOTE_DRIVER_VALIDATIONS = [
-    {'driver': 'gcs', 'path_regex': None},
-    {'driver': 's3', 'path_regex': None},
-]
-
 
 Index = types.Index
 Layout = layout.Layout
@@ -108,42 +102,6 @@ def get_tensorstore_spec(ckpt_path: str, ocdbt: bool = False):
       spec['kvstore'] = {'driver': ts_utils.DEFAULT_DRIVER, 'path': ckpt_path}
 
   return spec
-
-
-def is_remote_storage(tspec: Union[Dict[str, Any], str]) -> bool:
-  """Detect if user is using remote storages.
-
-  This can detect common defines and unable to detect some corner cases such as
-  using gcsfuse.
-
-  Args:
-    tspec: Tensorstore spec.
-
-  Returns:
-    True if the spec is using remote storage.
-  """
-  if isinstance(tspec, str):
-    # KvStoreUrl
-    if re.match(rf'^({"|".join(_REMOTE_URL_PREFIXES)})', tspec):
-      return True
-    else:
-      return False
-
-  for key in ('base', 'kvstore'):
-    if key in tspec:
-      return is_remote_storage(tspec[key])
-
-  if 'driver' in tspec:
-    for rule in _REMOTE_DRIVER_VALIDATIONS:
-      if tspec['driver'] == rule['driver']:
-        if rule['path_regex'] is None:
-          return True
-
-        # check if path matches the regex.
-        if re.match(rule['path_regex'], tspec['path']):
-          return True
-
-  return False
 
 
 class ByteLimiter(Protocol):