+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.5] - 2025-02-10
+
+### Fixed
+
+- Enable OCDBT read coalescing for remote storage
+
 ### Added
 
 - `ocp.metadata.get_step_metadata(path)` to public api.